( Library / Base Project / Run-Time ) Push 0.2 - Stable
+ Additionally Added  .GLM library (Included within include folder)
++ Add ( Stdafx.h file, which helps keep libraries pre-compiled avoiding complications of long build time / runtime )
The change requires that the libraries that don't significantly change throughout the project and do not need recompiling each end.

The idea I found significantly listed within the project with the included header, that did not change and having a read through to header.

My answer to the improvement / setup came from an old stackoverflow topic, which stated the significance of compile time within c++ apis, and the reason to set this header up is to reduce the compile time for each run as headers included there do not need to be compiled again.
More information can be found here:
<https://stackoverflow.com/questions/4726155/what-is-stdafx-h-used-for-in-visual-studio>

+ Base project
++ Established an understanding of each function clearly writing it out.
Improvements, in readability.

+ Run-Time is dependant on "stdafx.h " library.
+ Started a sphere class for Task 2.

- Problem Rectangle colours not represented correctly.!