Field of View FIx  Push 0.7
Though I am supposed to change the attributes of PixelCamera to define the lens of the camera, I instead decided to move the camera within the scene further away from the Objects.

Using RayOrigin as the camera location instead, as I am not certain how I would be required to define what I am asked of.