Background Addition Push 1.3
+ Additionally added a background, through a lot of pain and suffering I managed to figure it out.

Essentially we define the lower left corner of the screen
Then the Horizontal Line and Vertical line from it

And then define how the colours of the background will be displayed.

After it adds these calculated values to the final pixels that are considered empty compared to the Primitives spawned in the world.